+++ src/main/java/com/rcextract/minecord/ServerManager.java
@@ -1,7 +1,8 @@
 package com.rcextract.minecord;
 
 import java.util.Set;
-import java.util.UUID;
+
+import org.bukkit.OfflinePlayer;
 
 /**
  * A management system for Minecord Servers. This class only allows one instance to exist in the 
@@ -31,7 +32,7 @@ public interface ServerManager {
 	 * @param player The parameter of members of the target server.
 	 * @return The server the player is inside, regardless of its online state. Null if not found.
 	 */
-	public Server getServer(UUID player);
+	public Server getServer(OfflinePlayer player);
 	/**
 	 * Gets a server by if the user is inside.
 	 * @param user The parameter of members of the target server.
@@ -44,6 +45,12 @@ public interface ServerManager {
 	 * @return The server the channel is inside. Null if not found.
 	 */
 	public Server getServer(Channel channel);
+	/**
+	 * Gets the default server in use. All players outside a server will be automatically assigned 
+	 * to this server.
+	 * @return The default server.
+	 */
+	public Server getMain();
 	/**
 	 * Creates a server.
 	 * @param name The name of the server. Null value is not allowed.