+++ src/main/java/com/rcextract/minecord/User.java
@@ -1,11 +1,11 @@
 package com.rcextract.minecord;
 
 import java.util.List;
-import java.util.UUID;
 
-import org.bukkit.Bukkit;
+import org.bukkit.OfflinePlayer;
+import org.bukkit.entity.Player;
+import org.bukkit.permissions.Permission;
 
-import com.rcextract.minecord.event.ChannelSwitchEvent;
 import com.rcextract.minecord.event.UserEvent;
 
 public class User implements RecordManager<UserEvent> {
@@ -14,17 +14,17 @@ public class User implements RecordManager<UserEvent> {
 	private String name;
 	private String nickname;
 	private String desc;
-	private UUID player;
+	private OfflinePlayer player;
 	private Channel channel;
 	private Rank rank;
-	protected User(int id, String name, String nickname, String desc, UUID player, Channel channel, Rank rank) {
+	protected User(int id, String name, String nickname, String desc, OfflinePlayer player, Channel channel, Rank rank) {
 		this.id = id;
 		this.name = name;
 		this.nickname = nickname == null ? name : nickname;
 		this.desc = desc;
 		this.player = player;
 		this.channel = channel;
-		this.setRank(rank);
+		this.rank = rank;
 	}
 	public int getIdentifier() {
 		return id;
@@ -50,34 +50,43 @@ public class User implements RecordManager<UserEvent> {
 	public Channel getChannel() {
 		return channel;
 	}
-	public UUID getUUID() {
+	public OfflinePlayer getPlayer() {
 		return player;
 	}
-	public void setUUID(UUID player) {
+	public void setPlayer(OfflinePlayer player) {
 		this.player = player;
+	}
+	public boolean isOnline() {
+		return player.isOnline();
+	}
+	public Player getOnlinePlayer() {
+		return player.getPlayer();
 	}
 	/**
 	 * Assigns a user to this channel. This will override the current location of the user, 
 	 * meaning the user will be removed from the previous channel, after the user successfully 
 	 * joined the channel, defined by the return value of this method.
 	 * @param user The target user.
-	 * @return If the process succeeded.
 	 */
-	public boolean setChannel(Channel channel) {
-		ChannelSwitchEvent event = new ChannelSwitchEvent(channel == null ? Minecord.getServerManager().getServer("default").getChannelManager().getMainChannel() : channel, this);
-		Bukkit.getPluginManager().callEvent(event);
-		if (!(event.isCancelled())) {
-			this.channel = event.getChannel();
-			return true;
-		}
-		return false;
+	public void setChannel(Channel channel) {
+		if (channel == null) channel = Minecord.getServerManager().getMain().getChannelManager().getMainChannel();
+		this.channel = channel;
+		Minecord.updateMessage(this, true);
 	}
 	public Rank getRank() {
 		return rank;
 	}
 	public void setRank(Rank rank) {
-		if (rank == null) rank = Minecord.getServerManager().getServer("default").getRankManager().getMain();
+		if (rank == null) rank = Minecord.getServerManager().getMain().getRankManager().getMain();
 		this.rank = rank;
+	}
+	public void clear() throws IllegalStateException {
+		if (!(player.isOnline())) throw new IllegalStateException();
+		for (int i = 0; i < 25; i++) player.getPlayer().sendMessage("");
+	}
+	public void applyRank() throws IllegalStateException {
+		for (Permission permission : rank.getPermissions()) 
+			Minecord.getPermissionManager().playerAdd(null, player, permission.getName());
 	}
 	@Override
 	public List<UserEvent> getRecords() {